requirements.txt
torch
flask
diffusers
transformers
safetensors
requests
Pillow
